requirements.txt
pandas
scikit-learn
matplotlib
mysql-connector-python
joblib
Flask